Simpler IPC API
